requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-oauth-toolkit==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
filters==1.3.2
filters-django==1.0.5
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.1
psycopg2==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
